# api.yaml
swagger: "2.0.0"
info:
  title: Test API Documentation
  description: Api Documentation for Test App Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:     
...
  /v1/category/getSubCategoryByCategoryID/{categoryID}/{offset}/{limit}:
    get:
      tags:
        - name: Test
      summary: Sub Category By category Id 
      description: Getting all the Sub Categories By Category ID
      parameters:
        - name: categoryID
          in: path
          type: string 
          required: true
        - name: offset
          in: path
          type: string 
          required: true
        - name: limit
          in: path
          type: string 
          required: true
      responses:
        200:
          description: OK
...
  /v1/category/deleteCategory/{categoryID}:
    delete:
      tags:
        - name: Test
      summary: Deletes Category
      description: APi will delete the category in the db
      parameters:
        - name: categoryID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/category/deleteSubCategory/{subCategoryID}:
    delete:
      tags:
        - name: Test
      summary: Deletes sub Category
      description: APi will delete the Sub Category in the db
      parameters:
        - name: subCategoryID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/category/createCategory:
    post:
      tags:
        - name: Test
      summary: Will Create Categories for the test
      description: will generate categories regarding to the test
      consumes:
        application/json
      produces:
        application/json
      parameters:
      - in: body
        name: User
        schema:
          type: object
          properties:
            nameCategory:
              type: string 
            descriptionCategory:
              type: string
      responses:
        200:
          description: OK
...
  /v1/category/createSubCategory/{categoryID}:
    post:
      tags:
        - name: Test
      summary: Will Create Categories for the test
      description: will generate categories regarding to the test
      consumes:
        multipart/form-data
      produces:
        application/json
      parameters:
      - name: categoryID
        in: path
        required: true
      - name: imageLogo
        in: formData
        type: file
        required: true
      - name: nameSubCategory
        in: formData
        type: string
        required: true
      - name: descriptionSubCategory
        in: formData
        type: string
        required: true     
      responses:
        200:
          description: OK
  /v1/category/editSubCategory/{subCategoryID}:
    put:
      tags:
        - name: Test
      summary: Edit Sub Category
      description: lorem
      consumes:
        multipart/form-data
      produces:
        application/json
      parameters:
        - name: subCategoryID
          in: path
          type: string
          required: true
        - name: imageLogo
          in: formData
          type: file
          required: false
        - name: subCategory
          in: body
          schema:
            type: object
            properties:
              nameSubCategory:
                type: string
              descriptionSubCategory:
                type: string
      responses:
        200:
          description: OK
  /v1/category/editCategory/{CategoryID}:
    put:
      tags:
        - name: Test
      summary: Edit Category
      description: lorem
      consumes:
        multipart/form-data
      produces:
        application/json
      parameters:
        - name: CategoryID
          in: path
          type: string
          required: true
        - name: subCategory
          in: body
          schema:
            type: object
            properties:
              nameCategory:
                type: string
              descriptionCategory:
                type: string
      responses:
        200:
          description: OK
...
  /v1/category/getAllCategories/{offset}/{limit}:
    get:
      tags:
        - name: Test
      summary: Get all Categories
      description: APi will get all the test Categories
      parameters:
        - name: offset
          in: path
          type: integer
          required: true
        - name: limit
          in: path
          type: integer
          required: true
      responses:
        200:
          description: OK
...